config: Ensure configuration options are honoured
Client timeouts can now be specified on a `frontend`
using the `clientTimeout` option, with an appropriate
HAProxy value.

`backend`s now need to include any appropriate `server`
options (as well as a `url`) to ensure that these are
included in the configuration (hardcoded defaults have
been removed, as well as hardcoded checks for 'known'
components). This mostly involves including a `check`
and `port` option for most backends. See tests for
more details.

Connects-to: #18
Change-type: minor
Signed-off-by: Heds Simons <heds@balena.io>