confd: Remove `confd` and move to Docker envvars
Whilst moving to multiple envvar configs works, confd
does not support dynamic variables.

This PR moves to using the `/etc/docker.env` file for
the `balena-haproxy` service.

As this service was not intended to be used with a
system utilising `etcd`, we do not suffer any
consequences from this.

Connects-to: #33
Change-type: patch
Signed-off-by: Heds Simons <heds@balena.io>